mypy-requirements.txt
# NOTE: this needs to be kept in sync with the "requires" list in pyproject.toml
typing_extensions>=3.10
mypy_extensions>=1.0.0
typed_ast>=1.4.0,<2; python_version<'3.8'
tomli>=1.1.0; python_version<'3.11'
automata-lib>=7.1.0